Add unit test for qphot.run()
This is a test that should have been written a long time ago, and that
would have saved us some time trying to figure out what was failing in
issue #23. Test coverage of qphot.run() is not yet complete, but with
this we can finally verify that PyRAF and IRAF's qphot are working as
expected when we try to do photometry.