Update calls to LEMONdB.get_star()
Fix the ValueError ("too many values to unpack") exception raised by
the juicer GUI when we attempt to open a database: LEMONdB.get_star()
returns an eight-element tuple, but our code expects a five-element
one. This should have been done in commit 4cdc774, but I absolutely
forgot about it. Thanks to Scott Engle for reporting this (#33).