Clear the urlretrieve() cache in get_dss_image()
urllib.urlretrieve() does not copy the object if a valid cached copy
of the object exists. This causes get_dss_image() not to download an
image from the STScI Digitized Sky Survey if it has been previously
downloaded — for example, in two consecutive calls to the function.
Fix it by clearing with urllib.urlcleanup() the cache that may have
been built up by previous calls to urlretrieve(), so that we can
download an image for the same coordinates as many times as needed.