Fix LEMONdBTest.test_add_and_get_image_None_fields() (#41)
The message of the sqlite3.IntegrityError exception seems to have
changed between SQLite versions 3.7.13 (Debian Wheezy) and 3.8.7.1
(Jessie), as seen in #41. Modify test_add_and_get_image_None_fields()
so that assertRaisesRegexp() works with all versions of SQLite — the
exact wording of the exception is irrelevant, as long as it mentions
the name of the NULL attribute.