Ignore the photometric filter of the images unless --filter is given
The help text of the --filterk option states that the photometric
filter of the FITS images is not read from their headers unless the
--filter option is given. This contradicts the actual behavior, that
always raises KeyError if the keyword is not present, even if we do
not need it. Fix the code to not read it unless --filter is given.
Kudos to Pablo Ramírez for spotting this (#46).