Add workaround for bug in montage.mosaic() — background_match=True
There is a known bug in montage.mosaic() that raises an error if both
the 'mpi' and 'background_match' keyword arguments are set to True. As
a temporary workaround, set --cores to 1, automatically disabling MPI,
when the --background-match option is given by the user.

mpirun process exiting without calling "finalize":
https://github.com/astropy/montage-wrapper/issues/18