Fix bug in LEMONdB._add_cmp_star() — standard deviation and weight
The last two values in the INSERT INTO statement were in the wrong
order, and therefore the standard deviation of each comparison star
was stored as its weight in the artificial comparison star, and vice
versa. It is a bit surprising that nobody seems to have noticed this
in all this time.