Fix bug in LightCurveTest.assertThatAreEqual()
And this is why the bug in 87e114e960 went unnoticed: the light curve
that the LEMONdB returned was not the same as the input one (standard
deviations of the comparison stars and their weights in the artificial
comparison star were swapped), but this method was comparing the input
curve with itself. A bug in one of the unit tests — what a paradox.