Use 'travis_wait' to install requirements.txt — avoid timeout
The Travis build is timing out because SciPy takes too long to compile
without producing any output. Fix this with 'travis_wait', a function
that periodically writes a short line to the build log for 20 minutes,
extending the amount of time the 'pip' command has to finish.

How can I install scipy from pip on travis-ci.org without a timeout?
https://stackoverflow.com/q/28746046/184363