Move apt-get and pip commands to the .travis.yml file
I have just realized that the Travis documentation states that "There
is no timeout for a build; a build will run as long as all the jobs do
as long as each job does not timeout". The 50-minute timeout limit is
for each individual script or test suite — as it is our case with the
travis-setup.sh script. Move the apt-get and pip commands from there
to the .travis.yml file, so that each command is run with a separate
timeout.