Call mosaic() with BITPIX = -64
Emilio Falco reported in issue #60 a strange problem with Montage when
mosaicking one of his data sets (531 images): mConvert raises an
exception ("error reading from FITS file") when it attempts to change
the datatype of the intermediate FITS image from BITPIX = -64 (double
precision floating point) to -32 (single precision), as that is the
value used by default by mosaic().

I cannot reproduce the error, but in one of the tests we did we saw
that mosaic() runs without an error for him when 'bitpix' is set to
-64. Use from now on this as the BITPIX of the output FITS file: it
will make no difference in the final result (apart from a few more
megabytes), but it may help other people not to encounter this error.