Link to an example on GitHub when a filter cannot be recognized
As Pablo Ramírez pointed out, it is not very useful to tell users to
manually add a filter to the [custom_filters] section if they do not
even know what the ~/.lemonrc file is. Link to an example on GitHub as
a temporary solution until it is properly added to the documentation.