Use a delta of 1e-03 for floating-point comparison
The unit tests are failing in the Travis build, mysteriously
complaining that array(4.674883867122548) != 4.6749296929182735 within
1e-07 delta. This is something that I cannot reproduce locally using
neither a 32-bit nor a 64-bit architecture, and the underlying cause
seems to be the same addressed in 3942046a6e5c. I'm not going to dig
further into this, as we expect to transition to Astropy's photutils
soon.