Add 'barycentric' keyword to date()
At the same time, read_barycentric_date() becomes private. In this manner, FITSImage.date() is the single entry point, branching to either _read_uncorrected_date() (for the regular code path) or _read_barycentric_date() to read Baycentric Julian Dates. This allows us to extend the existing interface to add support for the BJD_TDB FITS keyword without having to update the rest of the pipeline.