Add hand-picked list of 58 stars in WASP-10b
SExtractor detects 3,035 objects on which to do photometry in WASP10b-mosaic.fits. This number is much greater than the actual number of stars in the image; while this doesn't affect the generated light curves —as false positives are discarded in the successive iterations— it vastly increases the execution time given the quadratic time complexity of Broeg's algorithm. This causes the test on Travis to hit the 50m build timeout and fail:

https://docs.travis-ci.com/user/customizing-the-build/#build-timeouts

Instead of relying on SExtractor, use a manually curated selection of 58 stars for the aperture photometry (and subsequent light curve generation) via --coordinates.