Pin version of most packages
This matches most of the package versions that I had on my workstation at the Institute of Astrophysics of Andalusia (CSIC) back in 2015. This gives me a working LEMON installation on my current Arch Linux machine, so that I can start migrating the codebase to more recent versions of these dependencies.