Fix column used for the default sort order of the curve data points
By default we sort the light curve data points by the 2nd column, but that's the string representation of the timestamps. Use instead the first (and hidden) column, with the Unix time of the data point, so that they are correctly sorted chronologically.

Thanks to @akdiaz for the help debugging this issue.