chmod u+w the (copy of the) sources image
`photometry` works on a temporary copy of the sources image. However, one of the steps in wasp10b-download.sh is `chmod a-w -v *.fits` (to make sure none of the integration tests, which use the same data and may run in parallel, accidentally modify any of the FITS files), and shutil.copy2() preserves these permissions. This results in errors such as "IOError: [Errno 13] Permission denied: '/tmp/WASP10b-mosaic_zN4uNy.fits'" when `photometry` attempts to modify the temporary copy of the sources image.